Preload fonts
Instead of loading the font files sequentially after the CSS styles,
start loading them as soon as the <head> tag is parsed.

This reduces the overall delay of loading the customs fonts. But it
might cause a font file to be downloaded unnecessarily, on cases where
the page doesn't have text with that font. This is why the code font is
not preloaded; not every page has code.