Simplify how spacing for sidenotes is computed
The previous way of doing this made it very difficult to adjust the
.wrapper's padding without tweaking lots of other things related to
sidenote spacing.

Now that padding is completely independent from how the sidenote spacing
is calculated, AND the code should be easier to follow. At the very
least, it doesn't need that long comment explaining the percentages of
everything to add up to 1460px.

Speaking of 1460px, this is no longer hard-coded. Things are not
calculated from a base CSS property that indicates the width of the main
text.

All these changes are completely transparent. The visual layout looks
and works exactly the same as before.